sdxl-lightning/requirements.txt
accelerate==0.28.0
bentoml>=1.2.9
diffusers==0.27.2
pillow==10.2.0
torch==2.2.1
transformers==4.39.1
xformers==0.0.25
